% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zones.R
\docType{data}
\name{default_zone_params}
\alias{default_zone_params}
\title{Default zone geometry parameters}
\format{
An object of class \code{list} of length 4.
}
\usage{
default_zone_params
}
\description{
A named list with `box_front_x`, `six_yard_front_x`, `near_cut`, and
`far_cut`, all in Opta's 0-100 normalized pitch units.
}
\keyword{datasets}
