% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_second_ball_map}
\alias{plot_second_ball_map}
\title{Full-pitch scatter of second-ball contests}
\usage{
plot_second_ball_map(contests, title = NULL)
}
\arguments{
\item{contests}{A list of contest records (from
[find_second_ball_contests()]).}

\item{title}{Optional plot title.}
}
\value{
A ggplot object.
}
\description{
Full-pitch scatter of second-ball contests
}
