% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{set_pieces_to_df}
\alias{set_pieces_to_df}
\title{Convert a list of set-piece events to a data.frame}
\usage{
set_pieces_to_df(events)
}
\arguments{
\item{events}{A list of set-piece event records (from
[extract_set_pieces()]).}
}
\value{
A data.frame with one row per event.
}
\description{
Convert a list of set-piece events to a data.frame
}
