% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setpieces-package.R
\docType{package}
\name{setpieces-package}
\alias{setpieces}
\alias{setpieces-package}
\title{setpieces: Extract and Analyze Football Set-Piece Events}
\description{
Tools to parse Opta MA1-style football event data and extract,
classify, and analyze set pieces -- corners, free kicks, throw-ins,
and penalties -- including delivery-zone breakdowns and second-ball
contest detection.
}
\author{
\strong{Maintainer}: Marc Lamberts \email{waltzinganalytics@gmail.com}

}
\keyword{internal}
